Move RadioListItem to its own file
